update: added custom hooks for escape and outside click handlers
