+++ lib/utils/useEscapeHandler.ts
@@ -1,0 +1,19 @@
+import { useEffect } from "react";
+
+const useEscapeHandler = (onClick: () => void) => {
+  useEffect(() => {
+    const handleEsc = (event: KeyboardEvent) => {
+      if (event.key === "Escape") {
+        onClick();
+      }
+    };
+
+    window.addEventListener("keydown", handleEsc);
+
+    return () => {
+      window.removeEventListener("keydown", handleEsc);
+    };
+  }, []);
+};
+
+export default useEscapeHandler;
